+++ Tasks/MoveTask.cpp
@@ -78,9 +78,9 @@ unsigned CMoveObjectTask::GetHelpResID()
 }
 
 
-unsigned long CMoveObjectTask::GetMask()
+DWORD CMoveObjectTask::GetMask()
 {
-	unsigned long mask = (
+	DWORD mask = (
 		TPoint3DClass | TPointSurD3DClass | TPointSurC3DClass | 
 		TPointSurP3DClass | TPointSurS3DClass | TPointSurCyl3DClass);
 
@@ -371,7 +371,7 @@ UINT CMoveObjectTask::GetTaskCursor()
 	return curID;
 }
 
-int	 CMoveObjectTask::FindObject(CPoint theLoc,unsigned long mask,BOOL bShowLabel,BOOL bSub)
+int	 CMoveObjectTask::FindObject(CPoint theLoc,DWORD mask,BOOL bShowLabel,BOOL bSub)
 {
 	return CTask::FindObject(theLoc,mask,bShowLabel,bSub);
 }
