+++ objects/ObjectPropPage.h
@@ -83,7 +83,7 @@ public:
 	CString		m_strVarName;
 	BOOL		m_bMarked;
 	//}}AFX_DATA
-	unsigned long	m_bMod;
+	DWORD 	m_bMod;
 	LOGFONT		m_lFont;
 	COLORREF	m_clrFont;
 	CMenu		m_menu;
